requirements.txt
fastapi
httpx
python-multipart
python-dotenv
